+++ traffic_portal/app/src/common/modules/form/server/FormServerController.js
@@ -60,6 +60,14 @@ var FormServerController = function(server, $scope, $location, $state, $uibModal
             });
     };
 
+    $scope.getProfileID = function(profileName) {
+        for (let profile of $scope.profiles) {
+            if (profile.name == profileName) {
+                return "/#!/profiles/"+profile.id
+            }
+        }
+    };
+
     var updateStatus = function(status) {
         serverService.updateStatus(server.id, { status: status.id, offlineReason: status.offlineReason })
             .then(
@@ -217,6 +225,9 @@ var FormServerController = function(server, $scope, $location, $state, $uibModal
         getTypes();
         getCDNs();
         getProfiles(($scope.server.cdnId) ? $scope.server.cdnId : 0); // hacky but does the job. only when a cdn is selected can we fetch the appropriate profiles. otherwise, show no profiles.
+        if ($scope.server.profileNames != undefined) {
+            $scope.server.profileName = $scope.server.profileNames[0]
+        }
     };
     init();