+++ traffic_portal/app/src/common/modules/form/server/edit/FormEditServerController.js
@@ -45,6 +45,7 @@ var FormEditServerController = function(server, $anchorScroll, $scope, $controll
     };
 
     $scope.save = function(server) {
+        console.log(server.profileNames = [server.profileName]);
         serverService.updateServer(server).
             then(
                 function(result) {